+++ typings/index.d.ts
@@ -4,7 +4,7 @@ interface TrimWhitespaceSettings {
 	AutoTrimDocument: boolean;
 	AutoTrimTimeout: number;
 
-	SkipCodeBlocks: boolean;
+	PreserveCodeBlocks: boolean;
 	PreserveIndentedLists: boolean;
 
 	TrimTrailingSpaces: boolean;